+++ src/app/App.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { ParallaxProvider } from 'react-scroll-parallax';
 import SMNavbar from '../components/layout/Navbar/SMNavbar';
 import Home from '../components/pages/Home/Home';
-import About from '../components/pages/AboutUsPage/AboutUsPage';
+import About from '../components/pages/AboutUsPage/SMAboutUsPage';
 import Services from '../components/pages/Services/Services';
 import ContactMe from '../components/pages/ContactMe/ContactMe';
 import './App.css';
